refactor: RacingGameConsole controller 패키지로 이동
